Test cluster sharding behaviors of all concert actor implementations
