+++ sample-app/src/test/scala/example/model/concert/actor/DefaultConcertActorWithEventPersistenceSpec.scala
@@ -1,11 +1,15 @@
 package example.model.concert.actor
 
 import akka.actor._
+import com.typesafe.config.ConfigFactory
 import example.ActorSpecBase
 
 final class DefaultConcertActorWithEventPersistenceSpec
-    extends ActorSpecBase(ActorSystem("default-concert-actor-with-event-persistence"))
-    with ConcertActorBehaviors {
+    extends ActorSpecBase(
+      ActorSystem("default-concert-actor-with-event-persistence", ConfigFactory.load("test-akka-cluster")),
+    )
+    with ConcertActorBehaviors
+    with ConcertActorClusterShardingBehaviors {
 
   private def createBehavior: ConcertActorBehaviorFactory = DefaultConcertActorWithEventPersistence
 
@@ -13,6 +17,7 @@ final class DefaultConcertActorWithEventPersistenceSpec
     behave like emptyConcertActor(new EmptyConcertActorFactory(createBehavior))
     behave like availableConcertActor(new AvailableConcertActorFactory(createBehavior))
     behave like cancelledConcertActor(new CancelledConcertActorFactory(createBehavior))
+    behave like shardedActor(createBehavior)
   }
 
 }
