+++ sample-app/src/test/scala/example/model/concert/actor/DefaultConcertActorWoPersistenceSpec.scala
@@ -1,11 +1,13 @@
 package example.model.concert.actor
 
 import akka.actor._
+import com.typesafe.config.ConfigFactory
 import example.ActorSpecBase
 
 final class DefaultConcertActorWoPersistenceSpec
-    extends ActorSpecBase(ActorSystem("default-concert-actor-wo-persistence"))
-    with ConcertActorBehaviors {
+    extends ActorSpecBase(ActorSystem("default-concert-actor-wo-persistence", ConfigFactory.load("test-akka-cluster")))
+    with ConcertActorBehaviors
+    with ConcertActorClusterShardingBehaviors {
 
   private def createBehavior: ConcertActorBehaviorFactory = DefaultConcertActorWoPersistence
 
@@ -13,6 +15,7 @@ final class DefaultConcertActorWoPersistenceSpec
     behave like emptyConcertActor(new EmptyConcertActorFactory(createBehavior))
     behave like availableConcertActor(new AvailableConcertActorFactory(createBehavior))
     behave like cancelledConcertActor(new CancelledConcertActorFactory(createBehavior))
+    behave like shardedActor(createBehavior)
   }
 
 }
